now the secret file will be automatically merged into the config file.
The secret file  must have the same structure and is then automatically merged into the config. The data types must match and only the fields listed in the secret file will be transferred/overwritten. In the case of an error, the start of the service is aborted with a corresponding error message.